+++ lib/binance_client/responses/sub_account_bnb_burn_status_response.rb
@@ -1,0 +1,17 @@
+module BinanceClient
+  class SubAccountBnbBurnStatusResponse < BaseResponse
+
+    def sub_account_id
+      body["subAccountId"]
+    end
+
+    def spot_bnb_burn?
+      body["spotBNBBurn"]
+    end
+
+    def interest_bnb_burn?
+      body["interestBNBBurn"]
+    end
+
+  end
+end